+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -10,5 +10,5 @@ public interface UserStorage {
 
     List<User> getAll();
 
-    Optional<User> findUserById(Integer id);
+    Optional<User> findUserById(Long id);
 }
